source/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
pyserial==3.5
Werkzeug==2.3.3
zipp==3.15.0
